requirements.txt
requests
beautifulsoup4
python-dotenv
pytest
newsapi-python
httpx
aiofiles